requirements.txt
scrapy>=2.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0.1
python-dateutil>=2.8.2
